+++ server/routes/cardRoutes.js
@@ -3,28 +3,13 @@
  */
 
 const express = require('express');
-const cardRouter = express.Router();
 const cardController = require('../controllers/cardControllers');
 
-// /card/ routes post requests to '/' to cardController
+const cardRouter = express.Router();
+
+// directs post requests made to the root endpoint of /card to the cardController
 cardRouter.post('/', cardController.addCard, (req, res) => {
   res.status(200).send(res.locals.newCard);
 });
-
-// [ ] hypothesis ... don't need a colon with params ...
-// /deck/[number]
-cardRouter.get('/deck/:deckId', cardController.readDeckOfCards, (req, res) => {
-  /*
-    from axios: https://oscards/deck/1
-    processed in express: req.params = {
-      deckId: 1
-    }
-  */
-  res.status(200).send(res.locals.data);
-});
-
-// cardRouter.get('/', cardController.readCard, (req, res) => {
-//   res.status(200).send(res.locals.data);
-// });
 
 module.exports = cardRouter;
