+++ server/server.js
@@ -7,13 +7,14 @@ const mongoose = require('mongoose');
 // invocation of dot env to access API keys
 require('dotenv').config();
 
-// creation of express instance 
+// creation of express instance
 const app = express();
 const PORT = 3000;
 const mongoURI = `${process.env.MONGO_URI}`;
 
 // connect to instance of mongodb atlas
-mongoose.connect(mongoURI, { useNewUrlParser: true, useUnifiedTopology: true})
+mongoose
+  .connect(mongoURI, { useNewUrlParser: true, useUnifiedTopology: true })
   .then(() => {
     app.listen(PORT, () => console.log(`Listening on PORT: ${PORT}`));
   })
@@ -28,22 +29,23 @@ if (process.env.NODE_ENV === 'production') {
   app.use('/dist', express.static(path.join(__dirname, '../dist')));
 }
 
-// instantiate router(s) for data calls 
+// instantiate router(s) for data calls
 const cardRouter = require('./routes/cardRoutes.js');
+const deckRouter = require('./routes/deckRoutes.js');
+
 app.use('/card', cardRouter);
+app.use('/deck', deckRouter);
 
 // serve index.html on the route '/'
-app.get('/', (req, res) => {
-  return res.status(200).sendFile(path.join(__dirname, '../client/src/index.html'));
-});
+app.get('/', (req, res) =>
+  res.status(200).sendFile(path.join(__dirname, '../client/src/index.html'))
+);
 
-// 404 error
-app.use('*', (req, res) => {
-  res.status(404).send('Not Found');
-});
+// create 404 error
+app.use('*', (req, res) => res.status(404).send('Not Found'));
 
-// global error handler
-app.use((err, req, res, next) => {
+// create global error handler
+app.use((err, req, res) => {
   console.log(err);
   res.status(500).send('Internal Server Error');
 });
